Fixed libtree-sitter only working in CWD by copying the lib to CWD
There is a store file keeping track of there the lib last was. During
boot, the IDE checks if the lib is in CWD. If it is, read it from there.
If it isn't, but the store file is pointing to a lib somewhere else,
remove the lib in the other directory, copy the lib to CWD, and update
store file to reflect that. If the store file doesn't exist, just copy
the file.